Move tests to its own module
