+++ modulo/src/lib.rs
@@ -4,5 +4,7 @@
 extern crate log;
 extern crate modulo_traits;
 
+#[cfg(test)]
+pub mod tests;
 pub mod core;
 pub mod file;
