+++ modulo/src/tests/file/mod.rs
@@ -1,0 +1,1 @@
+pub mod file_data;
